# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects with user progress
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject with chapters
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subject detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectDetail"
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Chapters
...
  /chapters/{chapterId}/topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List topics in a chapter with gate status
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a topic (admin only)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Created topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{topicId}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get topic detail with gate status
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Topic detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicDetail"
    patch:
      operationId: updateTopic
      tags: [topics]
      summary: Update topic (admin only)
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          description: Updated topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{topicId}/lecture-click:
    post:
      operationId: recordLectureClick
      tags: [topics]
      summary: Record a lecture click event
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Click recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Gate check
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List today's study tasks
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyTask"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a manual study task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyTask"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status or order
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyTaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyTask"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a study task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth schemas
...
    Subject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        iconName:
          type: ["string", "null"]
        totalChapters:
          type: integer
        completedChapters:
          type: integer
        progressPercent:
          type: number
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - order
        - totalChapters
        - completedChapters
        - progressPercent
        - gateStatus
    SubjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
        iconName:
          type: string
      required:
        - name
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
        iconName:
          type: string
    SubjectDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        iconName:
          type: ["string", "null"]
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        progressPercent:
          type: number
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
      required:
        - id
        - name
        - order
        - gateStatus
        - progressPercent
        - chapters

    # Chapters
    Chapter:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        totalTopics:
          type: integer
        completedTopics:
          type: integer
        progressPercent:
          type: number
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        chapterTestUnlocked:
          type: boolean
        notesUploadUnlocked:
          type: boolean
      required:
        - id
        - subjectId
        - name
        - order
        - totalTopics
        - completedTopics
        - progressPercent
        - gateStatus
        - chapterTestUnlocked
        - notesUploadUnlocked
...
    Topic:
      type: object
      properties:
        id:
          type: string
        chapterId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        telegramChatId:
          type: ["string", "null"]
        telegramMessageId:
          type: ["string", "null"]
        lectureQuizPassed:
          type: boolean
        dppCompleted:
          type: boolean
        pyqCompleted:
          type: boolean
        topicTestPassed:
          type: boolean
        isComplete:
          type: boolean
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        lectureClickCount:
          type: integer
      required:
        - id
        - chapterId
        - name
        - order
        - lectureQuizPassed
        - dppCompleted
        - pyqCompleted
        - topicTestPassed
        - isComplete
        - gateStatus
        - lectureClickCount
    TopicInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
        telegramChatId:
          type: string
        telegramMessageId:
          type: string
      required:
        - name
    TopicUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
        telegramChatId:
          type: string
        telegramMessageId:
          type: string
    TopicDetail:
      type: object
      properties:
        id:
          type: string
        chapterId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        telegramChatId:
          type: ["string", "null"]
        telegramMessageId:
          type: ["string", "null"]
        lectureQuizPassed:
          type: boolean
        dppCompleted:
          type: boolean
        pyqCompleted:
          type: boolean
        topicTestPassed:
          type: boolean
        isComplete:
          type: boolean
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        lectureClickCount:
          type: integer
        availableExams:
          type: array
          items:
            $ref: "#/components/schemas/Exam"
      required:
        - id
        - chapterId
        - name
        - order
        - lectureQuizPassed
        - dppCompleted
        - pyqCompleted
        - topicTestPassed
        - isComplete
        - gateStatus
        - lectureClickCount
        - availableExams

    # Gate
...
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [lecture_quiz, dpp, pyq, topic_test, chapter_test, subject_test, grand_test]
        subjectId:
          type: ["string", "null"]
        chapterId:
          type: ["string", "null"]
        topicId:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        totalQuestions:
          type: integer
        totalMarks:
          type: integer
        passingScore:
          type: ["integer", "null"]
        negativeMarking:
          type: number
        isUnlocked:
          type: boolean
        hasAttempted:
          type: boolean
        lastScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - durationMinutes
        - totalQuestions
        - totalMarks
        - negativeMarking
        - isUnlocked
        - hasAttempted
...
    StudyTask:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        source:
          type: string
          enum: [auto, manual]
        topicId:
          type: ["string", "null"]
        topicName:
          type: ["string", "null"]
        isLocked:
          type: boolean
        sortOrder:
          type: integer
        scheduledDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - status
        - source
        - isLocked
        - sortOrder
        - scheduledDate
    StudyTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        topicId:
          type: string
        scheduledDate:
          type: string
          format: date
      required:
        - title
    StudyTaskUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        sortOrder:
          type: integer
        title:
          type: string

    # Dashboard
...