# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EdTech Study Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  name?: string;
  description?: string;
  order?: number;
  iconName?: string;
}